1. Retrofit StartupNationalCloud to have most of services configured from the base Startup 2. Add appsettings per cloud environment
