+++ ApplensBackend/appsettings.Fairfax.json
@@ -72,5 +72,10 @@
   },
   "LocationPlacementIdService": {
     "Enabled": false
+  },
+  "DatacenterFederationConfiguration": {
+    "Realm": "",
+    "MetadataAddress": "",
+    "Issuer": ""
   }
 }
