+++ ApplensBackend/appsettings.Mooncake.json
@@ -72,5 +72,10 @@
   },
   "LocationPlacementIdService": {
     "Enabled": false
+  },
+  "DatacenterFederationConfiguration": {
+    "Realm": "",
+    "MetadataAddress": "",
+    "Issuer": ""
   }
 }
