+++ ApplensBackend/appsettings.USSec.json
@@ -72,5 +72,10 @@
   },
   "LocationPlacementIdService": {
     "Enabled": false
+  },
+  "DatacenterFederationConfiguration": {
+    "Realm": "",
+    "MetadataAddress": "",
+    "Issuer": ""
   }
 }
